# apps/visualizer/backend-ts/tests/fixtures/flask-openapi.json
{
...
 "components": {
  "schemas": {
...
   "CatalogCacheReadyResponse.36cf89b": {
    "additionalProperties": false,
    "properties": {
     "cached": {
      "title": "Cached",
      "type": "boolean"
     }
    },
    "required": [
     "cached"
    ],
    "title": "CatalogCacheReadyResponse",
    "type": "object"
   },
...
   "CatalogMonthsResponse.573ec44": {
    "additionalProperties": false,
    "properties": {
     "months": {
      "items": {
       "type": "string"
      },
      "title": "Months",
      "type": "array"
     }
    },
    "required": [
     "months"
    ],
    "title": "CatalogMonthsResponse",
    "type": "object"
   },
...
   "CatalogSimilarityGroupsResponse.573ec44": {
    "additionalProperties": false,
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44.CatalogSimilarityGroup"
      },
      "title": "Items",
      "type": "array"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total"
    ],
    "title": "CatalogSimilarityGroupsResponse",
    "type": "object"
   },
   "CatalogSimilarityGroupsResponse.573ec44.CatalogImage": {
    "additionalProperties": false,
    "description": "Catalog list / search row shape (``query_catalog_images`` + API transforms).",
    "properties": {
     "ai_analyzed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Ai Analyzed"
     },
     "analyzed_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Analyzed At"
     },
     "aperture": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Aperture"
     },
     "camera_make": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Camera Make"
     },
     "camera_model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Camera Model"
     },
     "caption": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Caption"
     },
     "catalog_path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Catalog Path"
     },
     "catalog_perspective_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Catalog Perspective Score"
     },
     "catalog_score_perspective": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Catalog Score Perspective"
     },
     "color_label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Color Label"
     },
     "copyright": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Copyright"
     },
     "date_taken": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Date Taken"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Description"
     },
     "description_best_perspective": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Description Best Perspective"
     },
     "description_summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Description Summary"
     },
     "exif": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Exif"
     },
     "file_size": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "File Size"
     },
     "filename": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Filename"
     },
     "filepath": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Filepath"
     },
     "focal_length": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Focal Length"
     },
     "gps_latitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Gps Latitude"
     },
     "gps_longitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Gps Longitude"
     },
     "height": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Height"
     },
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Id"
     },
     "image_hash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Image Hash"
     },
     "image_type": {
      "anyOf": [
       {
        "const": "catalog",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Image Type"
     },
     "instagram_posted": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Instagram Posted"
     },
     "is_stack_representative": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Is Stack Representative"
     },
     "iso": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Iso"
     },
     "key": {
      "title": "Key",
      "type": "string"
     },
     "keywords": {
      "items": {
       "type": "string"
      },
      "title": "Keywords",
      "type": "array"
     },
     "lens": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Lens"
     },
     "phash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Phash"
     },
     "pick": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Pick"
     },
     "rating": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Rating"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Score"
     },
     "shutter_speed": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Shutter Speed"
     },
     "similarity": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Similarity"
     },
     "stack_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Stack Id"
     },
     "stack_member_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Stack Member Count"
     },
     "thumbnail_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Thumbnail Url"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Title"
     },
     "why_matched": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Why Matched"
     },
     "width": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Width"
     }
    },
    "required": [
     "key"
    ],
    "title": "CatalogImage",
    "type": "object"
   },
   "CatalogSimilarityGroupsResponse.573ec44.CatalogSimilarityGroup": {
    "additionalProperties": false,
    "properties": {
     "best_similarity": {
      "title": "Best Similarity",
      "type": "number"
     },
     "candidate_count": {
      "title": "Candidate Count",
      "type": "integer"
     },
     "candidates": {
      "items": {
       "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44.CatalogImage"
      },
      "title": "Candidates",
      "type": "array"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Created At"
     },
     "group_id": {
      "title": "Group Id",
      "type": "integer"
     },
     "job_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Job Id"
     },
     "seed": {
      "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44.CatalogImage"
     }
    },
    "required": [
     "group_id",
     "seed",
     "candidates",
     "candidate_count",
     "best_similarity"
    ],
    "title": "CatalogSimilarityGroup",
    "type": "object"
   },
...
   "ErrorBody.45d9b59": {
    "properties": {
     "code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Code"
     },
     "error": {
      "title": "Error",
      "type": "string"
     }
    },
    "required": [
     "error"
    ],
    "title": "ErrorBody",
    "type": "object"
   },
...
   "ImageView.573ec44": {
    "additionalProperties": false,
    "description": "``GET /api/images/<image_type>/<image_key>`` consolidated detail shape.",
    "properties": {
     "added_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Added At"
     },
     "ai_analyzed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Ai Analyzed"
     },
     "analyzed_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Analyzed At"
     },
     "aperture": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Aperture"
     },
     "available_score_perspectives": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Available Score Perspectives"
     },
     "camera_make": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Camera Make"
     },
     "camera_model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Camera Model"
     },
     "caption": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Caption"
     },
     "catalog_path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Catalog Path"
     },
     "catalog_perspective_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Catalog Perspective Score"
     },
     "catalog_score_perspective": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Catalog Score Perspective"
     },
     "color_label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Color Label"
     },
     "copyright": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Copyright"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Created At"
     },
     "date_folder": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Date Folder"
     },
     "date_taken": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Date Taken"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Description"
     },
     "description_best_perspective": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Description Best Perspective"
     },
     "description_summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Description Summary"
     },
     "exif": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Exif"
     },
     "exif_data": {
      "anyOf": [
       {},
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Exif Data"
     },
     "file_path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "File Path"
     },
     "file_size": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "File Size"
     },
     "filename": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Filename"
     },
     "filepath": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Filepath"
     },
     "focal_length": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Focal Length"
     },
     "gps_latitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Gps Latitude"
     },
     "gps_longitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Gps Longitude"
     },
     "height": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Height"
     },
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Id"
     },
     "identity_aggregate_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Identity Aggregate Score"
     },
     "identity_eligible": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Identity Eligible"
     },
     "identity_peak_percentile": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Identity Peak Percentile"
     },
     "identity_per_perspective": {
      "items": {
       "$ref": "#/components/schemas/ImageView.573ec44.IdentityPerPerspectiveScore"
      },
      "title": "Identity Per Perspective",
      "type": "array"
     },
     "identity_perspectives_covered": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Identity Perspectives Covered"
     },
     "image_hash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Image Hash"
     },
     "image_type": {
      "const": "catalog",
      "title": "Image Type",
      "type": "string"
     },
     "instagram_folder": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Instagram Folder"
     },
     "instagram_posted": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Instagram Posted"
     },
     "is_stack_representative": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Is Stack Representative"
     },
     "iso": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Iso"
     },
     "key": {
      "title": "Key",
      "type": "string"
     },
     "keywords": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Keywords"
     },
     "last_attempted_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Last Attempted At"
     },
     "lens": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Lens"
     },
     "local_path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Local Path"
     },
     "matched_catalog_key": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Matched Catalog Key"
     },
     "media_key": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Media Key"
     },
     "phash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Phash"
     },
     "pick": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Pick"
     },
     "post_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Post Url"
     },
     "processed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Processed"
     },
     "processed_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Processed At"
     },
     "rating": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Rating"
     },
     "shutter_speed": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Shutter Speed"
     },
     "source_folder": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Source Folder"
     },
     "stack_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Stack Id"
     },
     "stack_member_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Stack Member Count"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Title"
     },
     "vision_result": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Vision Result"
     },
     "vision_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Vision Score"
     },
     "width": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Width"
     }
    },
    "required": [
     "image_type",
     "key"
    ],
    "title": "ImageView",
    "type": "object"
   },
   "ImageView.573ec44.IdentityPerPerspectiveScore": {
    "additionalProperties": false,
    "properties": {
     "display_name": {
      "title": "Display Name",
      "type": "string"
     },
     "model_used": {
      "title": "Model Used",
      "type": "string"
     },
     "percentile": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Percentile"
     },
     "perspective_slug": {
      "title": "Perspective Slug",
      "type": "string"
     },
     "prompt_version": {
      "title": "Prompt Version",
      "type": "string"
     },
     "rationale_preview": {
      "title": "Rationale Preview",
      "type": "string"
     },
     "score": {
      "title": "Score",
      "type": "integer"
     },
     "scored_at": {
      "title": "Scored At",
      "type": "string"
     }
    },
    "required": [
     "perspective_slug",
     "display_name",
     "score",
     "prompt_version",
     "model_used",
     "scored_at",
     "rationale_preview"
    ],
    "title": "IdentityPerPerspectiveScore",
    "type": "object"
   },
...
   "Stats.36cf89b": {
    "additionalProperties": false,
    "properties": {
     "catalog_images": {
      "title": "Catalog Images",
      "type": "integer"
     },
     "db_path": {
      "title": "Db Path",
      "type": "string"
     },
     "posted_to_instagram": {
      "title": "Posted To Instagram",
      "type": "integer"
     }
    },
    "required": [
     "catalog_images",
     "posted_to_instagram",
     "db_path"
    ],
    "title": "Stats",
    "type": "object"
   },
...
   "ValidationError.6a07bef": {
    "description": "Model of a validation error response.",
    "items": {
     "$ref": "#/components/schemas/ValidationError.6a07bef.ValidationErrorElement"
    },
    "title": "ValidationError",
    "type": "array"
   },
   "ValidationError.6a07bef.ValidationErrorElement": {
    "description": "Model of a validation error response element.",
    "properties": {
     "ctx": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Error context"
     },
     "loc": {
      "items": {
       "type": "string"
      },
      "title": "Missing field name",
      "type": "array"
     },
     "msg": {
      "title": "Error message",
      "type": "string"
     },
     "type": {
      "title": "Error type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationErrorElement",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Lightroom Tagger Visualizer API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/catalog/status": {
   "get": {
    "description": "",
    "operationId": "get__api_catalog_status",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogCacheReadyResponse.36cf89b"
        }
       }
      },
      "description": "OK"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Internal Server Error"
     }
    },
    "summary": "Return whether the catalog vision cache has any prepared entries.",
    "tags": [
     "system"
    ]
   }
  },
...
  "/api/images/catalog-similarity-groups": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog-similarity-groups",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "Reviewable catalog visual similarity groups materialized by batch jobs.",
    "tags": [
     "images-catalog"
    ]
   }
  },
...
  "/api/images/catalog/months": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog_months",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogMonthsResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "Get available year-months from catalog images based on date_taken.",
    "tags": [
     "images-catalog"
    ]
   }
  },
  "/api/images/catalog/{image_key}": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog_{image_key}",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "image_key",
      "required": true,
      "schema": {
       "format": "path",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ImageView.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "Single catalog image detail for the consolidated image-view modal.",
    "tags": [
     "images-catalog"
    ]
   }
  },
...
  "/api/stats": {
   "get": {
    "description": "",
    "operationId": "get__api_stats",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Stats.36cf89b"
        }
       }
      },
      "description": "OK"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Internal Server Error"
     }
    },
    "summary": "get_stats <GET>",
    "tags": [
     "system"
    ]
   }
  },
...
 },
 "security": [],
...
}